Added Victory Point checker to add/check victory point to tribes.

+++ Game/Data/Tribe/ITribe.cs
@@ -22,7 +22,7 @@ namespace Game.Data.Tribe
 
         int DefensePoint { get; set; }
 
-        int VictoryPoint { get; set; }
+        decimal VictoryPoint { get; set; }
 
         string Description { get; set; }
